+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `dataUris.sizeRatioPercent` metric
 - Added `selectors.attributesUsage` metric
 - Added `attributesUsage` flag
+- Added `imports` metric
 
 ### Changed